# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Boxes System API",
    "description": "基于 Spring Boot 的物资与盒子管理系统 API 文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "tags": [
    { "name": "用户管理", "description": "用户登录与增删改查" },
    { "name": "物品管理", "description": "物品全生命周期管理" },
    { "name": "盒子管理", "description": "盒子设备与状态管理" },
    { "name": "情感标签管理", "description": "物品情感备注管理" },
    { "name": "错误修复管理", "description": "运行时错误记录与降级" },
    { "name": "AI聊天", "description": "AI 助手交互" }
  ],
  "paths": {
    "/api/users/login": {
      "post": {
        "tags": ["用户管理"],
        "summary": "用户登录",
        "description": "验证账号密码并返回 Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Map" } } }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": ["用户管理"],
        "summary": "创建用户",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserDO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_UserDTO" } } } }
        }
      },
      "get": {
        "tags": ["用户管理"],
        "summary": "获取所有用户",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_List_UserDTO" } } } }
        }
      }
    },
    "/api/users/page": {
      "get": {
        "tags": ["用户管理"],
        "summary": "分页查询用户",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "页码" },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "每页大小" },
          { "name": "username", "in": "query", "schema": { "type": "string" }, "description": "用户名(模糊)" },
          { "name": "phone", "in": "query", "schema": { "type": "string" }, "description": "手机号" }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_PageResult_UserDTO" } } } }
        }
      }
    },
    "/api/items/add": {
      "post": {
        "tags": ["物品管理"],
        "summary": "添加新物品",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ItemDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/items/update": {
      "put": {
        "tags": ["物品管理"],
        "summary": "更新物品信息",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ItemDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/items/move/{itemId}/to/{boxId}": {
      "put": {
        "tags": ["物品管理"],
        "summary": "移动物品",
        "parameters": [
          { "name": "itemId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "boxId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/items/list": {
      "get": {
        "tags": ["物品管理"],
        "summary": "查询物品列表(无分页)",
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "boxId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "itemCode", "in": "query", "schema": { "type": "string" } },
          { "name": "name", "in": "query", "schema": { "type": "string" }, "description": "匹配自动名称或手动名称" },
          { "name": "itemTag", "in": "query", "schema": { "type": "string" } },
          { "name": "isValid", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_List_ItemDTO" } } } }
        }
      }
    },
    "/api/items/page": {
      "get": {
        "tags": ["物品管理"],
        "summary": "分页查询物品",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "sortField", "in": "query", "schema": { "type": "string", "default": "createTime" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "default": "desc" } },
          { "name": "userId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "boxId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "itemCode", "in": "query", "schema": { "type": "string" } },
          { "name": "name", "in": "query", "schema": { "type": "string" } },
          { "name": "itemTag", "in": "query", "schema": { "type": "string" } },
          { "name": "isValid", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_PageResult_ItemDTO" } } } }
        }
      }
    },
    "/api/boxes/add": {
      "post": {
        "tags": ["盒子管理"],
        "summary": "新增盒子",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BoxDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/boxes/update": {
      "put": {
        "tags": ["盒子管理"],
        "summary": "修改盒子信息",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BoxDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/boxes/report/status": {
      "put": {
        "tags": ["盒子管理"],
        "summary": "上报盒子状态",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BoxDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/boxes/page": {
      "get": {
        "tags": ["盒子管理"],
        "summary": "分页查询盒子",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "sortField", "in": "query", "schema": { "type": "string", "default": "createTime" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "default": "desc" } },
          { "name": "userId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "boxCode", "in": "query", "schema": { "type": "string" } },
          { "name": "boxName", "in": "query", "schema": { "type": "string" } },
          { "name": "boxType", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_PageResult_BoxDTO" } } } }
        }
      }
    },
    "/api/emotions/add": {
      "post": {
        "tags": ["情感标签管理"],
        "summary": "添加情感标签",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmotionDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/emotions/page": {
      "get": {
        "tags": ["情感标签管理"],
        "summary": "分页查询情感标签",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "itemId", "in": "query", "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_PageResult_EmotionDTO" } } } }
        }
      }
    },
    "/api/error-recovery/record-error": {
      "post": {
        "tags": ["错误修复管理"],
        "summary": "记录运行时错误",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorLogDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Void" } } } }
        }
      }
    },
    "/api/error-recovery/trigger-fallback": {
      "post": {
        "tags": ["错误修复管理"],
        "summary": "触发优雅降级",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FallbackRequestDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Object" } } } }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": ["AI聊天"],
        "summary": "发起对话",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiChatRequestDTO" } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestResult_Object" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RestResult_Void": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "msg": { "type": "string", "example": "Success" },
          "data": { "type": "null" }
        }
      },
      "RestResult_Object": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "msg": { "type": "string", "example": "Success" },
          "data": { "type": "object" }
        }
      },
      "RestResult_Map": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "msg": { "type": "string", "example": "Success" },
          "data": { "type": "object", "additionalProperties": true }
        }
      },
      "RestResult_UserDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": { "$ref": "#/components/schemas/UserDTO" }
        }
      },
      "RestResult_List_UserDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/UserDTO" } }
        }
      },
      "RestResult_PageResult_UserDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "total": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/UserDTO" } }
            }
          }
        }
      },
      "RestResult_List_ItemDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ItemDTO" } }
        }
      },
      "RestResult_PageResult_ItemDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "total": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/ItemDTO" } }
            }
          }
        }
      },
      "RestResult_PageResult_BoxDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "total": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/BoxDTO" } }
            }
          }
        }
      },
      "RestResult_PageResult_EmotionDTO": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "msg": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "total": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/EmotionDTO" } }
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["userAccount", "userPassword"],
        "properties": {
          "userAccount": { "type": "string", "description": "登录账号" },
          "userPassword": { "type": "string", "description": "登录密码" }
        }
      },
      "UserDO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userAccount": { "type": "string" },
          "userPassword": { "type": "string" },
          "username": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "userAccount": { "type": "string" },
          "username": { "type": "string" },
          "phone": { "type": "string" },
          "token": { "type": "string" }
        }
      },
      "ItemDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "itemCode": { "type": "string" },
          "boxId": { "type": "integer", "format": "int64" },
          "autoRecognizeName": { "type": "string" },
          "manualEditName": { "type": "string" },
          "itemTag": { "type": "string" },
          "itemDesc": { "type": "string" },
          "putInTime": { "type": "string", "format": "date-time" },
          "expireTime": { "type": "string", "format": "date-time" },
          "isValid": { "type": "integer" }
        }
      },
      "BoxDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "boxCode": { "type": "string", "maxLength": 32 },
          "userId": { "type": "integer", "format": "int64" },
          "boxName": { "type": "string", "maxLength": 50 },
          "boxType": { "type": "integer", "description": "0-子盒，1-主盒" },
          "status": { "type": "integer", "description": "0-离线 1-在线 2-低电量 3-故障" },
          "rssi": { "type": "integer" },
          "battery": { "type": "integer" },
          "networkDelay": { "type": "integer" },
          "lastHeartbeatTime": { "type": "string", "format": "date-time" },
          "createTime": { "type": "string", "format": "date-time" },
          "updateTime": { "type": "string", "format": "date-time" }
        }
      },
      "EmotionDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "itemId": { "type": "integer", "format": "int64" },
          "emotionTag": { "type": "integer", "description": "1-快乐 2-悲伤 3-愤怒 4-焦虑 5-平静" },
          "emotionRemark": { "type": "string", "maxLength": 500 }
        }
      },
      "ErrorLogDTO": {
        "type": "object",
        "properties": {
          "errorCode": { "type": "string" },
          "errorMessage": { "type": "string" },
          "errorTime": { "type": "string", "format": "date-time" }
        }
      },
      "FallbackRequestDTO": {
        "type": "object",
        "properties": {
          "serviceId": { "type": "string" },
          "fallbackReason": { "type": "string" }
        }
      },
      "AiChatRequestDTO": {
        "type": "object",
        "required": ["userId", "message"],
        "properties": {
          "userId": { "type": "integer", "format": "int64" },
          "message": { "type": "string", "maxLength": 1000 }
        }
      }
    }
  }
}